services/todo-grpc/requirements.txt
grpcio
grpcio-tools
protobuf
sqlalchemy
psycopg2-binary